@@ -15,17 +15,16 @@ public class ComplexFormatter : IFormatProvider, ICustomFormatter
     public string Format(string format, object arg,
                          IFormatProvider provider)
     {
-        if (arg is Complex)
+        if (arg is Complex c1)
         {
-            Complex c1 = (Complex)arg;
             // Check if the format string has a precision specifier.
             int precision;
-            string fmtString = String.Empty;
+            string fmtString = string.Empty;
             if (format.Length > 1)
             {
                 try
                 {
-                    precision = Int32.Parse(format.Substring(1));
+                    precision = int.Parse(format.Substring(1));
                 }
                 catch (FormatException)
                 {
@@ -42,12 +41,12 @@ public class ComplexFormatter : IFormatProvider, ICustomFormatter
         }
         else
         {
-            if (arg is IFormattable)
-                return ((IFormattable)arg).ToString(format, provider);
+            if (arg is IFormattable formattable)
+                return formattable.ToString(format, provider);
             else if (arg != null)
                 return arg.ToString();
             else
-                return String.Empty;
+                return string.Empty;
         }
     }
 }
@@ -56,22 +55,21 @@ public class ComplexFormatter : IFormatProvider, ICustomFormatter
 // <Snippet4>
 public class CustomFormatEx
 {
-    public static void Main()
+    public static void Run()
     {
-        Complex c1 = new Complex(12.1, 15.4);
-        Console.WriteLine("Formatting with ToString():       " +
-                          c1.ToString());
-        Console.WriteLine("Formatting with ToString(format): " +
-                          c1.ToString("N2"));
-        Console.WriteLine("Custom formatting with I0:        " +
-                          String.Format(new ComplexFormatter(), "{0:I0}", c1));
-        Console.WriteLine("Custom formatting with J3:        " +
-                          String.Format(new ComplexFormatter(), "{0:J3}", c1));
+        Complex c1 = new(12.1, 15.4);
+        Console.WriteLine($"Formatting with ToString:         {c1}");
+        Console.WriteLine($"Formatting with ToString(format): {c1:N2}");
+        Console.WriteLine($"Custom formatting with I0:\t" +
+            $"  {string.Format(new ComplexFormatter(), "{0:I0}", c1)}");
+        Console.WriteLine($"Custom formatting with J3:\t" +
+            $"  {string.Format(new ComplexFormatter(), "{0:J3}", c1)}");
     }
 }
+
 // The example displays the following output:
-//    Formatting with ToString():       (12.1, 15.4)
-//    Formatting with ToString(format): (12.10, 15.40)
+//    Formatting with ToString():       <12.1; 15.4>
+//    Formatting with ToString(format): <12.10; 15.40>
 //    Custom formatting with I0:        12 + 15i
 //    Custom formatting with J3:        12.100 + 15.400j
 // </Snippet4>
