@@ -20,18 +20,19 @@ class Program
     {
         // <Snippet1>
         // Create a span over an array.
-        var array = new byte[100];
-        var arraySpan = new Span<byte>(array);
+        byte[] array = new byte[100];
+        Span<byte> arraySpan = new(array);
 
         byte data = 0;
         for (int ctr = 0; ctr < arraySpan.Length; ctr++)
             arraySpan[ctr] = data++;
 
         int arraySum = 0;
-        foreach (var value in array)
+        foreach (byte value in array)
             arraySum += value;
 
         Console.WriteLine($"The sum is {arraySum}");
+
         // Output:  The sum is 4950
         // </Snippet1>
     }
@@ -40,7 +41,7 @@ class Program
     {
         // <Snippet2>
         // Create a span from native memory.
-        var native = Marshal.AllocHGlobal(100);
+        nint native = Marshal.AllocHGlobal(100);
         Span<byte> nativeSpan;
         unsafe
         {
@@ -51,11 +52,12 @@ class Program
             nativeSpan[ctr] = data++;
 
         int nativeSum = 0;
-        foreach (var value in nativeSpan)
+        foreach (byte value in nativeSpan)
             nativeSum += value;
 
         Console.WriteLine($"The sum is {nativeSum}");
         Marshal.FreeHGlobal(native);
+
         // Output:  The sum is 4950
         // </Snippet2>
     }
@@ -70,10 +72,11 @@ class Program
             stackSpan[ctr] = data++;
 
         int stackSum = 0;
-        foreach (var value in stackSpan)
+        foreach (byte value in stackSpan)
             stackSum += value;
 
         Console.WriteLine($"The sum is {stackSum}");
+
         // Output:  The sum is 4950
         // </Snippet3>
     }
@@ -85,8 +88,8 @@ public class ProgramB
     public static void WorkWithSpans()
     {
         // Create a span over an array.
-        var array = new byte[100];
-        var arraySpan = new Span<byte>(array);
+        byte[] array = new byte[100];
+        Span<byte> arraySpan = new(array);
 
         InitializeSpan(arraySpan);
         Console.WriteLine($"The sum is {ComputeSum(arraySpan):N0}");
@@ -118,14 +121,15 @@ public class ProgramB
             span[ctr] = value++;
     }
 
-    public static int ComputeSum(Span<byte> span)
+    public static int ComputeSum(ReadOnlySpan<byte> span)
     {
         int sum = 0;
-        foreach (var value in span)
+        foreach (byte value in span)
             sum += value;
 
         return sum;
     }
+
     // The example displays the following output:
     //    The sum is 4,950
     //    The sum is 4,950
