@@ -2,29 +2,50 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
+using Newtonsoft.Json;
 
 namespace CustomConverterUnixEpochDate
 {
     class Program
     {
-        public static void Main()
+        public static void STJExample(DateTimeOffset date)
         {
             var forecast = new Forecast()
             {
-                Date = DateTimeOffset.Now,
+                Date = date,
                 TemperatureCelsius = 19,
                 Summary = "warm"
             };
 
             var options = new JsonSerializerOptions();
             options.Converters.Add(new UnixEpochDateTimeOffsetConverter());
-            options.WriteIndented = true;
 
-            string json = JsonSerializer.Serialize(forecast, options);
-            Console.WriteLine(json);
+            string json = System.Text.Json.JsonSerializer.Serialize(forecast, options);
+            Console.WriteLine($"System.Text.Json: {json}");
 
-            Forecast forecastDeserialized = JsonSerializer.Deserialize<Forecast>(json, options)!;
-            Console.WriteLine($"Deserialized date = {forecastDeserialized.Date}");
+            Forecast forecastDeserialized = System.Text.Json.JsonSerializer.Deserialize<Forecast>(json, options)!;
+            Console.WriteLine($"System.Text.Json deserialized date = {forecastDeserialized.Date}");
+        }
+
+        public static void NewtonsoftExample(DateTimeOffset date)
+        {
+            var forecast = new Forecast()
+            {
+                Date = date,
+                TemperatureCelsius = 19,
+                Summary = "warm"
+            };
+
+            var settings = new JsonSerializerSettings
+            {
+                DateFormatHandling = DateFormatHandling.MicrosoftDateFormat,
+            };
+
+            string json = JsonConvert.SerializeObject(forecast, settings);
+            Console.WriteLine($"{Environment.NewLine}Newtonsoft: {json}");
+
+            Forecast forecastDeserialized = JsonConvert.DeserializeObject<Forecast>(json, settings)!;
+            Console.WriteLine($"Newtonsoft deserialized date = {forecastDeserialized.Date}");
         }
     }
 
@@ -36,23 +57,28 @@ namespace CustomConverterUnixEpochDate
     }
 
     // <ConverterOnly>
-    sealed class UnixEpochDateTimeOffsetConverter : JsonConverter<DateTimeOffset>
+    sealed class UnixEpochDateTimeOffsetConverter : System.Text.Json.Serialization.JsonConverter<DateTimeOffset>
     {
         static readonly DateTimeOffset s_epoch = new(1970, 1, 1, 0, 0, 0, TimeSpan.Zero);
-        static readonly Regex s_regex = new("^/Date\\(([+-]*\\d+)([+-])(\\d{2})(\\d{2})\\)/$", RegexOptions.CultureInvariant);
+        static readonly Regex s_regex = new(
+            "^/Date\\(([+-]*\\d+)([+-])(\\d{2})(\\d{2})\\)/$",
+            RegexOptions.CultureInvariant);
 
-        public override DateTimeOffset Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+        public override DateTimeOffset Read(
+            ref Utf8JsonReader reader,
+            Type typeToConvert,
+            JsonSerializerOptions options)
         {
             string formatted = reader.GetString()!;
             Match match = s_regex.Match(formatted);
 
             if (
                     !match.Success
-                    || !long.TryParse(match.Groups[1].Value, System.Globalization.NumberStyles.Integer, CultureInfo.InvariantCulture, out long unixTime)
-                    || !int.TryParse(match.Groups[3].Value, System.Globalization.NumberStyles.Integer, CultureInfo.InvariantCulture, out int hours)
-                    || !int.TryParse(match.Groups[4].Value, System.Globalization.NumberStyles.Integer, CultureInfo.InvariantCulture, out int minutes))
+                    || !long.TryParse(match.Groups[1].Value, NumberStyles.Integer, CultureInfo.InvariantCulture, out long unixTime)
+                    || !int.TryParse(match.Groups[3].Value, NumberStyles.Integer, CultureInfo.InvariantCulture, out int hours)
+                    || !int.TryParse(match.Groups[4].Value, NumberStyles.Integer, CultureInfo.InvariantCulture, out int minutes))
             {
-                throw new JsonException();
+                throw new System.Text.Json.JsonException();
             }
 
             int sign = match.Groups[2].Value[0] == '+' ? 1 : -1;
@@ -61,12 +87,18 @@ namespace CustomConverterUnixEpochDate
             return s_epoch.AddMilliseconds(unixTime).ToOffset(utcOffset);
         }
 
-        public override void Write(Utf8JsonWriter writer, DateTimeOffset value, JsonSerializerOptions options)
+        public override void Write(
+            Utf8JsonWriter writer,
+            DateTimeOffset value,
+            JsonSerializerOptions options)
         {
-            long unixTime = Convert.ToInt64((value - s_epoch).TotalMilliseconds);
+            long unixTime = value.ToUnixTimeMilliseconds();
+
             TimeSpan utcOffset = value.Offset;
 
-            string formatted = string.Create(CultureInfo.InvariantCulture, $"/Date({unixTime}{(utcOffset >= TimeSpan.Zero ? "+" : "-")}{utcOffset:hhmm})/");
+            string formatted = string.Create(
+                CultureInfo.InvariantCulture,
+                $"/Date({unixTime}{(utcOffset >= TimeSpan.Zero ? "+" : "-")}{utcOffset:hhmm})/");
 
             writer.WriteStringValue(formatted);
         }
