@@ -2,17 +2,20 @@
 using Microsoft.Extensions.AI;
 using OpenAI;
 
+// <SnippetGetChatClient>
 var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
 string model = config["ModelName"];
 string key = config["OpenAIKey"];
 
 // Create the IChatClient
 IChatClient chatClient =
-    new OpenAIClient(key).AsChatClient(model);
+    new OpenAIClient(key).GetChatClient(model).AsIChatClient();
+// </SnippetGetChatClient>
 
+// <SnippetFirstMessage>
 // Start the conversation with context for the AI model
-List<ChatMessage> chatHistory = new()
-    {
+List<ChatMessage> chatHistory =
+    [
         new ChatMessage(ChatRole.System, """
             You are a friendly hiking enthusiast who helps people discover fun hikes in their area.
             You introduce yourself when first saying hello.
@@ -27,19 +30,22 @@ List<ChatMessage> chatHistory = new()
             the local nature on the hikes when making a recommendation. At the end of your
             response, ask if there is anything else you can help with.
         """)
-    };
+    ];
+// </SnippetFirstMessage>
 
+// <SnippetChatLoop>
+// Loop to get user input and stream AI response
 while (true)
 {
     // Get user prompt and add to chat history
     Console.WriteLine("Your prompt:");
-    var userPrompt = Console.ReadLine();
+    string? userPrompt = Console.ReadLine();
     chatHistory.Add(new ChatMessage(ChatRole.User, userPrompt));
 
     // Stream the AI response and add to chat history
     Console.WriteLine("AI Response:");
-    var response = "";
-    await foreach (var item in
+    string response = "";
+    await foreach (ChatResponseUpdate item in
         chatClient.GetStreamingResponseAsync(chatHistory))
     {
         Console.Write(item.Text);
@@ -48,3 +54,4 @@ while (true)
     chatHistory.Add(new ChatMessage(ChatRole.Assistant, response));
     Console.WriteLine();
 }
+// </SnippetChatLoop>
