@@ -4,85 +4,85 @@ using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.VectorData;
 using Microsoft.SemanticKernel.Connectors.InMemory;
+using System.Linq;
 using VectorDataAI;
 
-var cloudServices = new List<CloudService>()
-{
-    new CloudService
-        {
-            Key=0,
-            Name="Azure App Service",
-            Description="Host .NET, Java, Node.js, and Python web applications and APIs in a fully managed Azure service. You only need to deploy your code to Azure. Azure takes care of all the infrastructure management like high availability, load balancing, and autoscaling."
-        },
-    new CloudService
-        {
-            Key=1,
-            Name="Azure Service Bus",
-            Description="A fully managed enterprise message broker supporting both point to point and publish-subscribe integrations. It's ideal for building decoupled applications, queue-based load leveling, or facilitating communication between microservices."
-        },
-    new CloudService
-        {
-            Key=2,
-            Name="Azure Blob Storage",
-            Description="Azure Blob Storage allows your applications to store and retrieve files in the cloud. Azure Storage is highly scalable to store massive amounts of data and data is stored redundantly to ensure high availability."
-        },
-    new CloudService
-        {
-            Key=3,
-            Name="Microsoft Entra ID",
-            Description="Manage user identities and control access to your apps, data, and resources.."
-        },
-    new CloudService
-        {
-            Key=4,
-            Name="Azure Key Vault",
-            Description="Store and access application secrets like connection strings and API keys in an encrypted vault with restricted access to make sure your secrets and your application aren't compromised."
-        },
-    new CloudService
-        {
-            Key=5,
-            Name="Azure AI Search",
-            Description="Information retrieval at scale for traditional and conversational search applications, with security and options for AI enrichment and vectorization."
-        }
-};
+// <SnippetDataSet>
+List<CloudService> cloudServices =
+[
+    new() {
+            Key = 0,
+            Name = "Azure App Service",
+            Description = "Host .NET, Java, Node.js, and Python web applications and APIs in a fully managed Azure service. You only need to deploy your code to Azure. Azure takes care of all the infrastructure management like high availability, load balancing, and autoscaling."
+    },
+    new() {
+            Key = 1,
+            Name = "Azure Service Bus",
+            Description = "A fully managed enterprise message broker supporting both point to point and publish-subscribe integrations. It's ideal for building decoupled applications, queue-based load leveling, or facilitating communication between microservices."
+    },
+    new() {
+            Key = 2,
+            Name = "Azure Blob Storage",
+            Description = "Azure Blob Storage allows your applications to store and retrieve files in the cloud. Azure Storage is highly scalable to store massive amounts of data and data is stored redundantly to ensure high availability."
+    },
+    new() {
+            Key = 3,
+            Name = "Microsoft Entra ID",
+            Description = "Manage user identities and control access to your apps, data, and resources."
+    },
+    new() {
+            Key = 4,
+            Name = "Azure Key Vault",
+            Description = "Store and access application secrets like connection strings and API keys in an encrypted vault with restricted access to make sure your secrets and your application aren't compromised."
+    },
+    new() {
+            Key = 5,
+            Name = "Azure AI Search",
+            Description = "Information retrieval at scale for traditional and conversational search applications, with security and options for AI enrichment and vectorization."
+    }
+];
+// </SnippetDataSet>
 
-// Load the configuration values
-var config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
+// <SnippetEmbeddingGenerator>
+// Load the configuration values.
+IConfigurationRoot config = new ConfigurationBuilder().AddUserSecrets<Program>().Build();
 string endpoint = config["AZURE_OPENAI_ENDPOINT"];
 string model = config["AZURE_OPENAI_GPT_NAME"];
 
-// Create the embedding generator
+// Create the embedding generator.
 IEmbeddingGenerator<string, Embedding<float>> generator =
-    new AzureOpenAIClient(
-        new Uri(endpoint),
-        new DefaultAzureCredential())
-            .AsEmbeddingGenerator(modelId: model);
+    new AzureOpenAIClient(new Uri(endpoint), new DefaultAzureCredential())
+        .GetEmbeddingClient(deploymentName: model)
+        .AsIEmbeddingGenerator();
+// </SnippetEmbeddingGenerator>
 
-// Create and populate the vector store
+// <SnippetVectorStore>
+// Create and populate the vector store.
 var vectorStore = new InMemoryVectorStore();
-Microsoft.Extensions.VectorData.IVectorStoreRecordCollection<int, CloudService> cloudServicesStore = vectorStore.GetCollection<int, CloudService>("cloudServices");
-await cloudServicesStore.CreateCollectionIfNotExistsAsync();
+VectorStoreCollection<int, CloudService> cloudServicesStore =
+    vectorStore.GetCollection<int, CloudService>("cloudServices");
+await cloudServicesStore.EnsureCollectionExistsAsync();
 
 foreach (CloudService service in cloudServices)
 {
-    service.Vector = await generator.GenerateEmbeddingVectorAsync(service.Description);
+    service.Vector = await generator.GenerateVectorAsync(service.Description);
     await cloudServicesStore.UpsertAsync(service);
 }
+// </SnippetVectorStore>
 
-// Convert a search query to a vector and search the vector store
+// <SnippetSearch>
+// Convert a search query to a vector
+// and search the vector store.
 string query = "Which Azure service should I use to store my Word documents?";
-ReadOnlyMemory<float> queryEmbedding = await generator.GenerateEmbeddingVectorAsync(query);
+ReadOnlyMemory<float> queryEmbedding = await generator.GenerateVectorAsync(query);
 
-VectorSearchResults<CloudService> results =
-    await cloudServicesStore.VectorizedSearchAsync(queryEmbedding, new VectorSearchOptions<CloudService>()
-{
-    Top = 1
-});
+IAsyncEnumerable<VectorSearchResult<CloudService>> results =
+    cloudServicesStore.SearchAsync(queryEmbedding, top: 1);
 
-await foreach (VectorSearchResult<CloudService> result in results.Results)
+await foreach (VectorSearchResult<CloudService> result in results)
 {
     Console.WriteLine($"Name: {result.Record.Name}");
     Console.WriteLine($"Description: {result.Record.Description}");
     Console.WriteLine($"Vector match score: {result.Score}");
-    Console.WriteLine();
 }
+// </SnippetSearch>
