@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.AI;
+﻿// <SnippetGetChatClient>
+using Microsoft.Extensions.AI;
 using Microsoft.Extensions.Configuration;
 using OpenAI;
 
@@ -7,10 +8,12 @@ string? model = config["ModelName"];
 string? key = config["OpenAIKey"];
 
 IChatClient client =
-    new ChatClientBuilder(new OpenAIClient(key).AsChatClient(model ?? "gpt-4o"))
+    new ChatClientBuilder(new OpenAIClient(key).GetChatClient(model ?? "gpt-4o").AsIChatClient())
     .UseFunctionInvocation()
     .Build();
+// </SnippetGetChatClient>
 
+// <SnippetAddOptions>
 // Add a new plugin with a local .NET function
 // that should be available to the AI model.
 var chatOptions = new ChatOptions
@@ -24,7 +27,9 @@ var chatOptions = new ChatOptions
     "get_current_weather",
     "Get the current weather in a given location")]
 };
+// </SnippetAddOptions>
 
+// <SnippetPromptModel>
 // System prompt to provide context.
 List<ChatMessage> chatHistory = [new(ChatRole.System, """
     You are a hiking enthusiast who helps people discover fun hikes in their area. You are upbeat and friendly.
@@ -37,3 +42,4 @@ Console.WriteLine($"{chatHistory.Last().Role} >>> {chatHistory.Last()}");
 
 ChatResponse response = await client.GetResponseAsync(chatHistory, chatOptions);
 Console.WriteLine($"Assistant >>> {response.Text}");
+// </SnippetPromptModel>
