@@ -1,7 +1,8 @@
 ﻿using Microsoft.Extensions.AI;
+using OllamaSharp;
 
 IChatClient chatClient =
-    new OllamaChatClient(new Uri("http://localhost:11434/"), "phi3:mini");
+    new OllamaApiClient(new Uri("http://localhost:11434/"), "phi3:mini");
 
 // Start the conversation with context for the AI model
 List<ChatMessage> chatHistory = new();
@@ -16,7 +17,7 @@ while (true)
     // Stream the AI response and add to chat history
     Console.WriteLine("AI Response:");
     var response = "";
-    await foreach (var item in
+    await foreach (ChatResponseUpdate item in
         chatClient.GetStreamingResponseAsync(chatHistory))
     {
         Console.Write(item.Text);
