@@ -1,4 +1,6 @@
-﻿namespace operators;
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace operators;
 
 public static class NullCoalescingOperator
 {
@@ -7,6 +9,35 @@ public static class NullCoalescingOperator
         WithNullConditional();
         WithNullableTypes();
         NullCoalescingAssignment();
+    }
+
+    public record class Human(string FirstName, string LastName)
+    {
+        public string FirstName { get; set; } = FirstName;
+    }
+    public static void AddMessageAtIndex()
+    {
+        List<string> messages = new List<string>(10);
+        Human person = new Human("First", "Last");
+        // <NullForgivingAssignment>
+        person?.FirstName = "Scott";
+        messages?[5] = "five";
+        // </NullForgivingAssignment>
+
+        int index = 0;
+        int[] values = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
+
+        // <ConditionalRHS>
+        values?[2] = GenerateNextIndex();
+        int GenerateNextIndex() => index++;
+        // </ConditionalRHS>
+
+        // <EquivalentIfStatement>
+        if (values is not null)
+        {
+            values[2] = GenerateNextIndex();
+        }
+        // </EquivalentIfStatement>
     }
 
     private static void WithNullConditional()
