@@ -5,15 +5,16 @@ class Program2
     static void Run()
     {
         string contentLength = "Content-Length: 132";
-        var length = GetContentLength(contentLength.ToCharArray());
+        int length = GetContentLength(contentLength.ToCharArray());
         Console.WriteLine($"Content length: {length}");
     }
 
     private static int GetContentLength(ReadOnlySpan<char> span)
     {
-        var slice = span.Slice(16);
+        ReadOnlySpan<char> slice = span.Slice(16);
         return int.Parse(slice);
     }
 }
+
 // Output:
 //      Content length: 132
