@@ -11,7 +11,7 @@ namespace CopyOptions
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             Forecast forecast = new()
             {
