@@ -101,7 +101,7 @@ namespace CustomConverterPreserveReferences
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             Employee tyler = new()
             {
