@@ -22,7 +22,7 @@ namespace Fields
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             string json = """
                 {
