@@ -17,7 +17,7 @@ namespace ImmutableTypesCtorParms
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             string json = """
                 {
