@@ -17,7 +17,7 @@ namespace JsonIgnoreAttributeExample
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             Forecast forecast = new()
             {
