@@ -16,7 +16,7 @@ namespace NonPublicAccessors
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             string json = """
                 {
