@@ -12,7 +12,7 @@ namespace PreserveReferences
 
     public class Program
     {
-        public static void Main()
+        public static void Run()
         {
             Employee tyler = new()
             {
