@@ -5,87 +5,91 @@
         static async Task Main(string[] args)
         {
             Console.WriteLine("======== Custom converter Unix Epoch DateTimeOffset =========");
-            CustomConverterUnixEpochDate.Program.Main();
+            DateTimeOffset date = DateTimeOffset.Now;
+            CustomConverterUnixEpochDate.Program.STJExample(date);
+            CustomConverterUnixEpochDate.Program.NewtonsoftExample(date);
             Console.WriteLine();
 
             Console.WriteLine("======== Custom converter Unix Epoch DateTime =========");
-            CustomConverterUnixEpochDateNoZone.Program.Main();
+            DateTime date2 = DateTime.Now;
+            CustomConverterUnixEpochDateNoZone.Program.STJExample(date2);
+            CustomConverterUnixEpochDateNoZone.Program.NewtonsoftExample(date2);
             Console.WriteLine();
 
             Console.WriteLine("======== Preserve references =========");
-            PreserveReferences.Program.Main();
+            PreserveReferences.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Custom converter Preserve references =========");
-            CustomConverterPreserveReferences.Program.Main();
+            CustomConverterPreserveReferences.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Preserve references Multiple calls =========");
-            PreserveReferencesMultipleCalls.Program.Main();
+            PreserveReferencesMultipleCalls.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Immutable types =========");
-            ImmutableTypes.Program.Main();
+            ImmutableTypes.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Immutable types =========");
-            ImmutableTypesCtorParms.Program.Main();
+            ImmutableTypesCtorParms.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Record support =========");
-            Records.Program.Main();
+            Records.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Field support =========");
-            Fields.Program.Main();
+            Fields.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Non-string key dictionary =========");
-            NonStringKeyDictionary.Program.Main();
+            NonStringKeyDictionary.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== HttpClient extension methods =========");
-            await HttpClientExtensionMethods.Program.Main();
+            await HttpClientExtensionMethods.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Custom converter handle null =========");
-            CustomConverterHandleNull.Program.Main();
+            CustomConverterHandleNull.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Custom converter inferred types to object =========");
-            CustomConverterInferredTypesToObject.Program.Main();
+            CustomConverterInferredTypesToObject.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Ignore value type default on serialize =========");
-            IgnoreValueDefaultOnSerialize.Program.Main();
+            IgnoreValueDefaultOnSerialize.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Ignore null on serialize =========");
-            IgnoreNullOnSerialize.Program.Main();
+            IgnoreNullOnSerialize.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Conditionally ignore selected properties on serialize =========");
-            JsonIgnoreAttributeExample.Program.Main();
+            JsonIgnoreAttributeExample.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Non-public accessors =========");
-            NonPublicAccessors.Program.Main();
+            NonPublicAccessors.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Copy options instance =========");
-            CopyOptions.Program.Main();
+            CopyOptions.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Create options instance with specified defaults =========");
-            OptionsDefaults.Program.Main();
+            OptionsDefaults.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== Quoted numbers =========");
-            QuotedNumbers.Program.Main();
+            QuotedNumbers.Program.Run();
             Console.WriteLine();
 
             Console.WriteLine("======== GuidReferenceResolver =========");
-            GuidReferenceResolverExample.Program.Main();
+            GuidReferenceResolverExample.Program.Run();
             Console.WriteLine();
         }
     }
