@@ -9,10 +9,10 @@ public class Example
       String fmt = (result < TimeSpan.Zero ?  "\\-" : "") + "dd\\.hh\\:mm";
 
       Console.WriteLine(result.ToString(fmt));
-      Console.WriteLine($"Interval: {result:" + fmt + "}");
+      Console.WriteLine($"Interval: {result.ToString(fmt)}");
    }
 }
 // The example displays output like the following:
-//       -1291.10:54
-//       Interval: -1291.10:54
+//       -5582.12:21
+//       Interval: -5582.12:21
 // </Snippet29>
