@@ -6,11 +6,11 @@ public partial class Program
     // Import user32.dll (containing the function we need) and define
     // the method corresponding to the native function.
     [LibraryImport("user32.dll", StringMarshalling = StringMarshalling.Utf16, SetLastError = true)]
-    private static partial int MessageBox(IntPtr hWnd, string lpText, string lpCaption, uint uType);
+    private static partial int MessageBoxW(IntPtr hWnd, string lpText, string lpCaption, uint uType);
 
     public static void Main(string[] args)
     {
         // Invoke the function as a regular managed method.
-        MessageBox(IntPtr.Zero, "Command-line message box", "Attention!", 0);
+        MessageBoxW(IntPtr.Zero, "Command-line message box", "Attention!", 0);
     }
 }
